Photometry.py: first step of function separation
Created a parallel portion of code that operates on a single aggregated table of photometry measurements. This also required an update to ComputeNetRateJupiter.py to add SessionID as a parameter. As of this commit, both threads of code produce identical, valid output. The next step will be to delete the old code and use only the new.